Move to RenderEffect commit_hash:53d759dc3b8e2ad5e4062474df4f17f62c93de56

+++ client/android/div/src/main/java/com/yandex/div/core/view2/divs/BaseDivViewExtensions.kt
@@ -863,7 +863,7 @@ internal fun View.applyBitmapFilters(
             when (filter) {
                 is DivFilter.Blur -> {
                     val radius = filter.value.radius.evaluate(resolver).toIntSafely().dpToPx(resources.displayMetrics)
-                    result = bitmapEffectHelper.blurBitmap(result, radius)
+                    result = bitmapEffectHelper.blurBitmap(result, radius.toFloat())
                 }
 
                 is DivFilter.RtlMirror -> if (isLayoutRtl()) {
